requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2
pymongo
openpyxl